requirements.txt
playwright
pandas
openpyxl
pyarrow
supabase
python-dotenv
